👁️
Very slight update needed to fix broken description text on the IWHI page (well, not broken, I just accidentally set it to scroll instead of letting it overflow. I also took the liberty of adding a Twitch link icon next to the new Instagram icon on the home page so I can give that a boost… and of course, I changed the footer to say "ALL RIGHTS RESERVED 👁️ AREAZONE.ORG".